move htmlToPlainText out of common/utils.ts
