+++ frontends/main/src/common/utils.test.ts
@@ -4,7 +4,6 @@ import {
   externalLinkProps,
   addExternalLinkTargets,
   stripAnchorTags,
-  htmlToPlainText,
 } from "./utils"
 
 const NEXT_PUBLIC_ORIGIN = process.env.NEXT_PUBLIC_ORIGIN
@@ -90,36 +89,6 @@ describe("stripAnchorTags", () => {
   it("leaves text with no anchors unchanged", () => {
     const html = stripAnchorTags("<p>Safe text</p>")
     expect(html).toBe("<p>Safe text</p>")
-  })
-})
-
-describe("htmlToPlainText", () => {
-  it("strips tags and decodes entities", () => {
-    expect(htmlToPlainText("<p>Daryl Morey &amp; Jessica Gelman</p>")).toBe(
-      "Daryl Morey & Jessica Gelman",
-    )
-  })
-
-  it("keeps a space between adjacent block-level elements", () => {
-    expect(htmlToPlainText("<p>First</p><p>Second</p>")).toBe("First Second")
-  })
-
-  it("keeps a space where a <br> separates lines", () => {
-    expect(htmlToPlainText("Line one<br>Line two")).toBe("Line one Line two")
-  })
-
-  it("strips links but keeps their text", () => {
-    expect(
-      htmlToPlainText('<p><a href="https://ocw.mit.edu">OCW</a> resources</p>'),
-    ).toBe("OCW resources")
-  })
-
-  it("leaves plain text unchanged", () => {
-    expect(htmlToPlainText("Just plain text")).toBe("Just plain text")
-  })
-
-  it("returns an empty string for empty input", () => {
-    expect(htmlToPlainText("")).toBe("")
   })
 })
